FilterWindow X 누르면 정상적으로 닫히게 만들기
